feat: introduce `maybe_init()`
With `maybe_init()` you can try to initialize a logger, but it will
silently fail in case it cannot be initialized. This is useful in
tests, where there might already be some logger running. It is more
important that some logging is happening (as opposed to logging with
this specific loggger), rather then failing due to the logger not
being able to be initialized.