fix: make local times work again (#19)
Due to https://rustsec.org/advisories/RUSTSEC-2020-0159, it wasn't possible
to get the local time in multi-threaded contexts. This was fixed with the
latest release v0.24 of `flexi_logger`. It is using a fixed version of
`chrono`, which also means that we don't have a dependency on `time` anymore.

Fixes #17.
Closes #18.